bix1_ops: add SetSleepInterval command

+++ rust/bix1_ops_app/src/rtc/config.rs
@@ -8,7 +8,7 @@ pub struct BixConfig {
     pub current_time: i64,
     pub next_reset: i64,
     pub reset_interval: u64,
-    pub seconds_to_wait_before_sleeping: u64
+    pub seconds_to_wait_before_sleeping: u32
 }
 
 impl Default for BixConfig {
